+++ src/client/mod.rs
@@ -25,3 +25,4 @@
 pub mod client;
 #[allow(unused_must_use)]
 pub mod client_config;
+pub mod raw_client;
